move Pipe and Socket demos into corresponding test files
