+++ ouroboros-network/test/Test/Pipe.hs
@@ -1,16 +1,31 @@
-{-# LANGUAGE DataKinds         #-}
-{-# LANGUAGE FlexibleInstances #-}
-{-# LANGUAGE TypeOperators     #-}
+{-# LANGUAGE ScopedTypeVariables #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 module Test.Pipe (tests) where
 
-import           Ouroboros.Network.Pipe (demo)
-
+import           Control.Monad
+import           Control.Monad.Class.MonadFork
+import           Control.Monad.Class.MonadSTM
+import           Control.Monad.Class.MonadTimer
+import qualified Data.ByteString.Lazy as BL
+import           System.Process (createPipe)
 import           Test.ChainGenerators (TestBlockChainAndUpdates (..))
-
 import           Test.QuickCheck
 import           Test.Tasty (TestTree, testGroup)
 import           Test.Tasty.QuickCheck (testProperty)
+
+import           Network.TypedProtocol.Driver
+
+import           Ouroboros.Network.Chain (Chain, ChainUpdate, Point)
+import qualified Ouroboros.Network.Chain as Chain
+import qualified Ouroboros.Network.ChainProducerState as CPS
+import           Ouroboros.Network.Channel
+import qualified Ouroboros.Network.Mux as Mx
+import           Ouroboros.Network.Pipe
+import           Ouroboros.Network.Protocol.ChainSync.Client
+import           Ouroboros.Network.Protocol.ChainSync.Codec
+import           Ouroboros.Network.Protocol.ChainSync.Examples
+import           Ouroboros.Network.Protocol.ChainSync.Server
+import           Ouroboros.Network.Serialise
 
 --
 -- The list of all tests
@@ -22,7 +37,6 @@ tests =
   [ testProperty "pipe sync demo"        prop_pipe_demo
   ]
 
-
 --
 -- Properties
 --
@@ -30,4 +44,93 @@ tests =
 prop_pipe_demo :: TestBlockChainAndUpdates -> Property
 prop_pipe_demo (TestBlockChainAndUpdates chain updates) =
     ioProperty $ demo chain updates
+
+
+-- | The enumeration of all mini-protocols in our demo protocol.
+data DemoProtocols = ChainSync
+  deriving (Eq, Ord, Enum, Bounded)
+
+instance Mx.ProtocolEnum DemoProtocols where
+  fromProtocolEnum ChainSync = 2
+
+  toProtocolEnum 2 = Just ChainSync
+  toProtocolEnum _ = Nothing
+
+-- | A demonstration that we can run the simple chain consumer protocol
+-- over a pipe with full message serialisation, framing etc.
+--
+demo :: forall block .
+        (Chain.HasHeader block, Serialise block, Eq block )
+     => Chain block -> [ChainUpdate block] -> IO Bool
+demo chain0 updates = do
+
+    (hndRead1, hndWrite1) <- createPipe
+    (hndRead2, hndWrite2) <- createPipe
+
+    producerVar <- atomically $ newTVar (CPS.initChainProducerState chain0)
+    consumerVar <- atomically $ newTVar chain0
+    consumerDone <- atomically newEmptyTMVar
+
+    let Just expectedChain = Chain.applyChainUpdates updates chain0
+        target = Chain.headPoint expectedChain
+        a_mps ChainSync = Mx.MiniProtocolDescription
+                            (Mx.AppProtocolId ChainSync)
+                            (consumerInit consumerDone target consumerVar)
+                            dummyCallback
+        b_mps ChainSync = Mx.MiniProtocolDescription
+                            (Mx.AppProtocolId ChainSync)
+                            dummyCallback
+                            (producerRsp producerVar)
+
+    startPipe b_mps (hndRead1, hndWrite2)
+    startPipe a_mps (hndRead2, hndWrite1)
+
+    void $ fork $ sequence_
+        [ do threadDelay 10000 -- just to provide interest
+             atomically $ do
+                 p <- readTVar producerVar
+                 let Just p' = CPS.applyChainUpdate update p
+                 writeTVar producerVar p'
+             | update <- updates
+        ]
+
+    atomically $ takeTMVar consumerDone
+
+  where
+    checkTip target consumerVar = atomically $ do
+          chain <- readTVar consumerVar
+          return (Chain.headPoint chain == target)
+
+    consumerClient :: Point block -> TVar IO (Chain block) -> Client block IO ()
+    consumerClient target consChain =
+      Client
+        { rollforward = \_ -> checkTip target consChain >>= \b ->
+            if b then pure $ Left ()
+                 else pure $ Right $ consumerClient target consChain
+        , rollbackward = \_ _ -> checkTip target consChain >>= \b ->
+            if b then pure $ Left ()
+                 else pure $ Right $ consumerClient target consChain
+        , points = \_ -> pure $ consumerClient target consChain
+        }
+
+    consumerInit :: TMVar IO Bool -> Point block -> TVar IO (Chain block)
+                 -> Channel IO BL.ByteString -> IO ()
+    consumerInit done_ target consChain channel = do
+       let consumerPeer = chainSyncClientPeer (chainSyncClientExample consChain
+                                               (consumerClient target consChain))
+
+       runPeer codecChainSync channel consumerPeer
+       atomically $ putTMVar done_ True
+
+       return ()
+
+    dummyCallback _ = forever $
+        threadDelay 1000000
+
+    producerRsp ::  TVar IO (CPS.ChainProducerState block)
+                -> Channel IO BL.ByteString -> IO ()
+    producerRsp prodChain channel = do
+        let producerPeer = chainSyncServerPeer (chainSyncServerExample () prodChain)
+
+        runPeer codecChainSync channel producerPeer
